[style] adjust access of dictionary elements
This assumes that if the key has a value, then
the value is a dict, but the current code makes
a similar assumption, however it would have
failed in a different way.

Also, these code blocks are quite similar, you
might consider refactoring and adding a free
function to do this particular task.