Use egrep to find repository components in arbitrary positions.
The existing implementation assumes that the repositories in question
are part of the sources.list on a line of their own when they may exist
as one component among several on the same line. This trades the fixed
string check for regular expression that will find the desired component
regardless of its position among multiple components or if it is on a
line by itself.