Use egrep to find repository components in arbitrary positions. (#532)
* Use egrep to find repository components in arbitrary positions.

The existing implementation assumes that the repositories in question
are part of the sources.list on a line of their own when they may exist
as one component among several on the same line. This trades the fixed
string check for regular expression that will find the desired component
regardless of its position among multiple components or if it is on a
line by itself.

* Allow dashes in component names.

Such as `non-free`...

* Escape . in patterns.

It's unlikely these patterns would appear with anything other than a
literal '.' in these files, and in fact there was an existing pattern in
a sed command which also used wildcard dots rather than quoted dots but
this replaces all wildcard dots in patterns with the expected quoted
dot.

* Remove doubled trailing slash and make trailing slash optional.

The doubled trailing slash appears to've been added in
c95cbb3a41e80c413b671f1fea5f7167a48ef997. Since there is no explanation
for it in the commit message I assume it was added unintentionally and
have removed it.

Additionally, a review of the apt line in our current osrf/ubuntu_arm*
images where ports.ubuntu.com is likely to appear has no trailing
slashes, which means we'd still end up with doubled apt repositories,
but perhaps becauase the urls are distinct we don't get warnings for it.
To guard against future changes, let's just make the trailing slash
optional for all patterns.