Add ddebs to published files.
ddebs containing debug symbols stripped from binary packages are being
generated by default for artful and bionic. Reprepro cannot ingest ddeb
files so we removed them from the changes file in reprepro updater (see
https://github.com/ros-infrastructure/reprepro-updater/pull/55).

Since we would like to provide debug symbols we've found a slightly
different approach: renaming the ddeb file to a regular deb (see
https://github.com/ros-infrastructure/reprepro-updater/pull/60).

In order for that to work, we actually need to publish the ddeb
to the repo host with the other files. I opted to publish it as
a ddeb and rename it at import time so that no changes need to be
made to ros_buildfarm if we change the handling in the future.