WIP: pass environment variables from release build to binarydeb task.
TODO: The release_build file itself is no longer explicitly known at
this point. I've started down a path of passing it along through the
entire ros_buildfarm stack but I'm pushing this WIP commit to see if
others with more experience in this codebase have a more convenient or
preferred way to access the desired information.