Initialize build_environment_variables to None.
Following the convetion of the list attributes I initialized this
variable previously to an empty dict. But feedback suggests initializing
to None is preferred. It is also more consistent with how the
build_environment_variables will be handled when passed as arguments to
release scripts as the argument parser initializes the collection to
None if no matching argument is passed.