Set build environment variables from build files. (#554)
* Add build_environment_variables dict to release build configuration.

* Add snippet for including environment variables in dockerfiles.

* Pass environment variables from release build to binarydeb task.

* Add argument parser for environment variables.

* Pass build file environment variables to binarydeb job.

Environment variables from the release build file are passed through the
job template to the run_binarydeb_job.py script which will in turn pass
them to the generated Dockerfile where they're used.

* Pass environment variables to the binarydeb docker generation script.

* Set build environment variables before dependencies are installed.

These build-time environment variables will be present in subsequent
layers of the docker image and in the final container image as well.

By setting the environment variables in the layers above dependency
installation it is possible for packages with environment-controlled
post-install scripts to be controlled using these variables.

* Initialize build_environment_variables to None.

Following the convetion of the list attributes I initialized this
variable previously to an empty dict. But feedback suggests initializing
to None is preferred. It is also more consistent with how the
build_environment_variables will be handled when passed as arguments to
release scripts as the argument parser initializes the collection to
None if no matching argument is passed.

* Sort template data by the order it is accessed in templates.

* Initialize build_environment_variables to an empty list.

* Pass --env-vars unconditionally.

The argument when passed with no values defaults to an empty list
whereas when not passed defaults to None.

* Document build_environment_variables option in release-build files.

* Move build_environment_variables to BuildFile class.

In order to let it be used for other job types, move the
build_environment_variables option to the BuildFile superclass.

* Add build_environment_variables to devel jobs.