Use globals to expose data values to templates
The way local scope is handle in Python 2 vs Python 3 differs, and
using locals to pass this data to EmPy results in unexpected behavior
when a new local scope is created, for example, when a generator
expression is used.

This affects not only the data we're passing to the template, but also
any values defined within the template itself, which can be very
difficult to work around.

Since we know the values of prior to processing, we can expose these
values as globals instead.