Use Bourne shell / dash compatible shell condition. (#592)
* Use Bourne shell / dash compatible shell condition.

The construct is used in dockerfile templates to retry failed apt
installations.

    RUN for i in 1 2 3; do apt-get update && apt-get install -q -y locales && apt-get clean && break || if [[ $i < 3 ]]; then sleep 5; else false; fi; done

This doesn't work in the /bin/sh implementation used by Debian and
Ubuntu Docker images which use dash, the Debian Almquist Shell [1].

The `[[ $i < 3 ]]` construct is not valid Bourne shell, but easy to miss
since it works in many current shell implementations. In Bourne shell
it's an attempt to run the command `[[ $i` with the file `3` directed to
its stdin and is failing to find the command `[[`.

The changed version uses the [ alias of the test command and the
arithmetic -lt less-than comparison. If a bash version is preferred we
could wrap the entire thing in bash -c '...' but if we do so we should
change ``[[ $i < 3 ]]` to a `(( $i < 3 ))` to ensure arithmetic comparison
rather than string comparison best illustrated by the following
examples:

    bash -c 'if [[ 10 < 3 ]]; then echo yikes; else echo phew; fi'
    bash -c 'if (( 10 < 3 )); then echo yikes; else echo phew; fi'

[1]: https://en.wikipedia.org/wiki/Almquist_shell#dash:_Ubuntu,_Debian_and_POSIX_compliance_of_Linux_distributions

* Apply the shell change to a further template.