Handle relocation of externally-built workspaces
The setup.sh can't reliably determine the location of the workspace, and
will assume that the workspace is in the same location as it was
initially built unless told otherwise.

This is a requirement of the build tool which built the workspace, not
necessarily the build tool being used in the current build, so we should
set the variables regardless of whether we're using catkin or colcon.