Add in the capability for a sync package percent.
That is, a sync will only happen if the percentage of
binary packages in a release is over a certain threshold.

All three of 'packages', 'package_count', and 'package_percent'
are ANDed together (if they are specified), meaning that a sync
will only happen if all of the conditions specified are met.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>