Fix CI build detection of non-ROS packages (#642)
During the dependency walk, any ROS packages encountered as a dependency 
were recursively walked, and therefore aren't included in the resulting
enumeration. Beyond that, we remove any keys which have been
specifically ignored (via `skip_rosdep_keys`). Another quirk was added
such that if a package is present but was specifically ignored, we also
exclude it from the dependencies to avoid installing a deb package to
satisfy the dependency.

The change is to take that a step further and remove dependencies on any 
non-ROS packages present in the workspaces. Instead of specifically
enumerating the non-ROS packages, I'm just removing any package (ROS,
non-ROS, ignored, etc) present in the workspace.

The new behavior should be that the only time a ROS deb is installed is 
if the package listed as a dependency isn't part of ANY of the
workspaces AND the package isn't specifically ignored via
`skip_rosdep_keys`.