Manually inspect colcon index to find CI underlay packages
Using colcon to enumerate packages in the underlay is not sufficient to
detect non-ROS packages which were installed there. However, this
information was captured in colcon's index when the workspace was built.

We don't need to worry about this for catkin-built workspaces because
all of the packages built there were ROS packages, which `catkin_pkg`
can detect.