Dirty support for multiple underlays in CI jobs
This change implements support for multiple underlays by laying each
underlay on top of the other in the same directory.

A better solution would be to source each underlay one at a time, but
that would mean that each underlay is mounted at a different path within
the container, and that is not currently supported since some packages
are not currently relocatable.

Signed-off-by: Scott K Logan <logans@cottsay.net>