Add support for Jenkins credential binding plugin
Right now, I'm authenticating with pulp using a username and password.
In order to store those credentials in Jenkins and make them accessible
to a script, I need to bind the credentials to environment variables so
that they don't get dumped to the logs as a command line argument.

This snippet takes a Jenkins credential ID and a prefix, and sets the
${PREFIX}_USERNAME and ${PREFIX}_PASSWORD environment variables for the
scripts to access.