Move debian-specific release templates to a subdirectory (#717)
Each of these templates includes a debian-specific approach that doesn't
apply to RPMs. The templates remaining in the `templates/release`
directory don't appear to be debian-specific and should be usable for
RPMs as well.

In some locations (such as `release_job.py`) it will be necessary to
select a template based on the packaging format of the target OS.
Organizing the templates in subdirectories like this will give us an
easy way to select the template that serves a specific purpose for a
given packaging format. This is why some of the template file names were
also renamed to be more neutral.