Add RPM upload script (#731)
This script uploads RPM packages to pulp. The uploaded packages are not yet associated with a
repository - that will happen during the package import step.

The pulp HREF of each uploaded package, which uniquely identifies the package on the pulp
instance, is written to a file so that it can be passed from the source or binary job to the import
job later on.