Include filtered packages in status pages
Previously, packages which were explicitly excluded from the release
builds by means of white/black-listing in the release build
configuration were completely excluded from the status page. Downstream
dependencies of those packages are, however, included.

This doesn't align with the Jenkins jobs, where jobs are created for
both explicitly and implicitly excluded packages, but are disabled.

This change leverages the already present "intentionally missing" state
on the status page to reflect that the package is part of the
distribution index, but isn't expected to be present in the build
results. It also exposes packages which should have been missing from
the build results but are present anyway.

This change doesn't handle the implicitly excluded packages, but sets a
precedent for how they could be displayed in a subsequent change.