Mark implicitly excluded packages in status pages
This change causes packages which are implicitly excluded by the build
configuration to be marked as 'intentionally missing'.

The explicitly excluded packages are already marked as such, but the
jobs for packages which depend on explicitly disabled packages are
disabled, and are therefore expected to be missing from the build
results as well.