Extract empty group dependencies for topological ordering.
This is a fix for an issue introduced by #758. Now that conditional
dependencies are being evaluated and checked, an assert is triggering
during release jobs since group memberships are not being extracted.

However, if we were to extract true group memberships we could create a
discrepancy between the topological order according to ros_buildfarm and
the dependency tree used by bloom as bloom does not currently support
group dependencies.