Never skip ros_buildfarm RPM repositories (#780)
For whatever reason, mock is skipping repositories if the metadata
download fails, which is not the default behavior according to YUM.

Interestingly enough, the system-provided mock configs explicitly set
this configuration to 'False' for all of the OS repositories. We should
do the same.

This will give us a more accurate fatal error - a repository metadata
error - instead of a cryptic "package not found".