Don't consider source package name if not provided
Evidently the 'Source:' value in a debian `Packages` list is not
required, so we might not know what source package corresponds to a
given binary package.

If we don't know for sure, list the package as 'missing' as we did
before instead of ignoring it.

Follow-up to 3a80a38363feb491160f458d083aa4af8ab9b7ce