Enable RPM debug packages (#806)
The one caveat to this approach is that when a binary package is removed
from the primary repository, it is not automatically removed from the
debug repository. However, this is already the situation for source and
binary packages, so I'm not concerned. It shouldn't cause any problems
if we end up with excess symbol packages in the repository.