Show additional context on Jenkins job diffs.
When trying to debug the differences between an old and new
job, a single line diff may not give enough context as to
what actually changed.  There's no exactly correct number to
use here, but I've increased it to 5 lines before and after
as a compromise between being too verbose and giving enough
context.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>