Disable automatic out-of-source CMake builds in RPMs
In Fedora 33, the CMake macro was changed to automatically create and
use a subdirectory for build files. We're already doing that in the
Bloom templates manually, so we need to specify a macro to retain the
existing behavior or the two approaches will collide and make the build
step fail.

Once the new behavior is reliably part of all Fedora and EPEL systems
we're targeting, we can consider modifying the Bloom templates to
leverage the new feature.