Add upload_destination_credential_id config option
This credential will be used to convey destination information
associated with `upload_credential_id`. In the case of Pulp, this will
specify the base URL to the API endpoint.

The 'credentials_binding_plugin' snippet is now deprecated in favor of
the 'credential_bindings' snippet, which follows the model used in the
'property_parameters-definition' snippet.