Abort release builds if already re-queued
There seems to be a race of some kind regarding the triggering of
downstream jobs and the execution of jobs already in the queue. What
I've observed is that if a build is in the queue but is waiting for a
single dependency to finish building, when the dependency finishes, the
build starts executing immediately. The dependency's build, however,
triggers another build of that same job that is immediately re-added to
the queue.

One theory is that there is some delay between when a new build is
"triggered" and when that build is actually enqueued. So the sequence
may be:
1. A is building, while B is waiting for A
2. A finishes building and requests a build of downstream project B
3. B is dequeued and begins execution
4. A's build request for B is processed and B is enqueued again

An alternative strategy to avoid this strange behavior might be to enact
some sort of quiet period AFTER a build triggers new downstream builds
but before the build is actually regarded as 'complete'.