Implement rosdoc2 documentation.
This will allow us to use https://github.com/ros-infrastructure/rosdoc2
to build API and other documentation for individual packages.

Most of this code was copied from the "rosdoc_lite" code and then
adapted as appropriate.

As it stands, these jobs have a few limitations (some of which
come from this code, some of which come from rosdoc2):

1.  This will really only work for C++ packages.  To make it work
    for Python requires an update to the rosdoc2 code.
2.  It will only generate documentation against sources.  To make
    it look at build or install artifacts as well, we'll need to
    update both rosdoc2 and this code.
3.  It will not do any cross-referencing outside of the current
    repository.  To make this work requires generating, storing,
    and retrieving a cross-reference database.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>