Update configuration files for new plugin versions.
Changes to plugin configuration

* hudson.tasks.Shell added the ability to filter or reset environment
variables from specific steps and the default configuration saves a
self-closing configuredLocalRules tag which contains that information.

* warnings-ng no longer uses a referenceJobName field and now has a more
complex/robust method for selecting a reference build using the Git
Forensics plugin
https://github.com/jenkinsci/warnings-ng-plugin/blob/master/doc/Documentation.md#configure-the-selection-of-the-reference-build-baseline

* warnings-ng has additional configuration for publishing checks to
GitHub. It is unclear if the successful use of this feature requires
usng the GitHub Branch Source plugin which we have not explored but the
if publishing checks works, we'll do it.
`publishAllIssues` is an additional configuration for publishing checks for
all issues not just issues which are `new` as of the current build.
For now I've left this as true until we have more information about the
reliability of the reference build info with our current setup.

* xunit report types can now take an exclude pattern as well as the
include pattern. At one point we split up some test type patterns to
deconflict between multiple test types but I'm not sure that a simple
exclude would allow us to re-combine any of them I will have to check
the history.

* xunit inner threshold tags are no longer required.  There are also
several new extra configuration fields for a sleep timer between reports (to avoid
report CPU time starving out the Jenkins ping thread) a stdout/stderr
log reduction option to save memory during report generation, and an
option whether or not to follow symlinks when reviewing reports. I have
opted to leave all of these at their defaults for now. There is
additionally a final `testDataPublishers` tag with the attribute
"empty-set" I cannot info on what this does but it appears to be a new
feature not an evolution of something we were already doing.