Add tag for followSymlinks default value.
As far as I can tell this default setting did not previously require
a tag but now does. I think it's also worth considering _not_ following
symlinks for the archiver in this snippet as we only ever intend to
archive the generated archive from this build and no other contents
outside the workspace.