Use prereq repos in sync job container (#888)
During the build for the container which is used to determine if the
sync criteria is met for a sync-to-testing job, that container is
currently provisioned with the repository it is checking. This change
instead provisions the container with the repositories listed as
"prerequisite" repositories in the build configuration index, and not
those associated with each build file.