Work around a packaging bug in RHEL 8.4 (#889)
There is a fix pending for this issue, but it is currently breaking
pretty much all builds on the farm, and a hack to the source RPM build
processes is the most expedient way to work around the bug.

This change can be reverted once the CentOS 8 buildroot has an updated
cmake package which provides the cmake3 virtual package. A rebuild of
the source packages will not be required at that time.