Add a clang-tidy warning publisher. (#897)
* Add a clang-tidy warning publisher.

This is most desired for CI jobs but I don't see a reason to restrict
deployment to just those jobs. We can always revisit that if this
warning parser picks up false positives in other jobs.

* Run clang-tidy parser on build output rather than log files.

Based on the PR conversation there's no advantage to running this
warning parser on the individual log files rather than the build log
output.