Drop "global" CI configuration (#915)
This was almost certainly an oversight when I implemented the CI jobs.
The CI jobs only work when they're associated with a distro, and this
section of the index is used for distro-independent configurations. We
aren't using this config, and it wouldn't do anything even if we were.

Best to simply drop it so that it isn't misleading, and add a comment to
clarify for future additions.