Bump pulp task timeout to 2 minutes
It seems that the import operations on the repo host can be somewhat CPU
constrained, and under load, the jobs can normally take 50 seconds to
complete. Even though we've taken steps to bring those times down on the
production buildfarm, 60 seconds is no longer a good place to declare a
task as 'stuck'.

Also update to the new model name. It seems to have changed in a
previously adopted release.