Separate objects when passing cache and new packages to pulp sync (#931)
Previously, the `package_cache` was duplicated in
`import_and_invalidate` before being modified. A recent change modified
the code to use a passed object as-is. Because the same instance was
also passed for `packages_to_add`, modifications to the cache resulted
in changes to the list of packages to add.

The result was that no packages were ever removed from the target repo,
effectively making the invalidation pattern inert.

Regression caused by 6343f5e771247ac797b6cffbc917382edd7e9b98