Update label expression to include agent_on_jenkins.
The agent_on_master was renamed to agent_on_jenkins in the migration to
cookbook-ros-buildfarm and Ubuntu 20.04. To maintain compatibility the
cookbook still applies the label agent_on_master to the agent_on_jenkins
but since Jenkins upstream is updating the name of the built-in agent
this seems like a good time to update the labels in ROS build farm.

A future release of ROS Buildfarm can remove the old names and simplify
the label expressions again.