Stop implicitly installing the ros_workspace package (#950)
This will only affect ROS 2 builds.

At some point, it was necessary for CI and/or Devel jobs to
unconditionally install the ros_workspace package. There are some recent
developments and goals which would benefit from discontinuing this
practice and letting the debians install the workspace as a dependency
only when called for.