Add support for bare file paths in index URL
This change enhances the URL resolution to support specifying bare file
paths for the index URL. Previously, absolute file paths could be
converted to URL syntax ('file:///foo/bar'), but with this change, any
specified URL which lacks a schema will be treated as a file path. A
relative file path is also supported, and is resolved from the current
working directory (as expected).

This change leverages these traits of 'urljoin':
* If the url lacks a schema and is relative, the two urls are combined
* If the url lacks a schema and is absolute, the base_url schema is
  combined with the original url
* If the url has a schema, the url is not changed