Fix the argument type on --pulp-task-timeout (#961)
Without specifying the type, it will default to `str`, which will result
in an error trying to add floats and strings. The type conversion
doesn't get applied when the default value is used, so I didn't notice
the problem until I actually tried to specify something on the command
line.