Fix some scripts which didn't previously require ros_buildfarm
It seems that these four scripts didn't previously require the
ros_buildfarm module to be importable to run. Rather than update all of
their calls throughout ros_buildfarm (there are numerous), take the same
approach as was used in the git and apt wrapper scripts previously to
inject the module into the import search path.

Fixes: 642dd451ec5c9c7c8813997134009a71d1b440cd