Add support for group dependencies in release jobs
Neither the deb nor the RPMs currently generated by Bloom support this,
but implementing the concept here is straightforward and it will make it
easier to prototype solutions if it can be enabled via the build config.

This change should maintain the current behavior on the buildfarm, which
ignores group dependencies in release jobs (except for topological
ordering).