Restore exec dependencies during non-test system packaging (#994)
The exec dependencies were dropped in a previous change to control the
inclusion of test dependencies. The exec dependencies must be present to
run tests, but because they aren't necessarily needed to build, they
were dropped when the tests aren't going to run. However, the
buildfarm's invalidation logic will remove downstream package which
declare an exec_depend on a package, so the relationship must be
maintained even though the invalidation isn't actually necessary.