Refactor groovy script to include batching.
The last several years, we've added batching to the this template script
when executing it on the official build farms but never actually updated
it here.

This change includes what is essentially, a rewrite of the batch process example script to more closely match the one I've been using the last year or so.

I someday hope to get even more meta about this and create a management
job on Jenkins that runs variations of this script on templated job types in order to remove the need for batching and the chance for a typo when copy-pasting and modifying the example script.