Deal with RPM *-SPECPARTS build subdirectory
In typical RPM builds, the `BUILD` directory contains only a single
subdirectory, which is where the package sources are extracted to prior
to building.

Beginning with RPM 4.19, a -SPECPARTS subdirectory is created to
facilitate dynamic spec file generation. We need to explicitly ignore it
when we go looking for where the sources were extracted.