Add jammy rosdoc2 jobs as well. (#1037)
I couldn't figure out if there was a better solution than just doubling
the existing job. I marked the previous one as for Noble and the one I
just added for Humble as "jammy" I guess that when we introduce new
platforms we'd roll these forward or back based on what is currently
supported.