Set PODMAN_USERNS=keep-id when invoking 'docker run' (#1032)
The way Dockerfiles are structured in ros_buildfarm, we're dependent on
the UID staying the same when the container is run.

Setting PODMAN_USERNS=keep-id when the `docker` command is backed by
Podman will in effect preserve the behavior of Docker we're relying on
here. When the `docker` command is backed by Docker, this shouldn't
change the behavior at all.