Replace `if not jenkins` with `if jenkins is False`.
There's a malinteraction with the jenkinsapi library and Python
convention which has been causing extremely large HTTP requests whenever
the truthiness of an instance of the Jenkins class is tested.

Jenkins defines a `__len__` method which is used to determine
truthiness. But this method triggers a fetch of every job on our Jenkins
host which can take 2 - 5 minutes to return due to the sheer volume of
jobs.

By avoiding this in favor of an explicit check of `is False` or `is
None` as appropriate we can save taking a really big runtime hit and cut
down on spurious failures due to timeouts.