Use None consistently as sigil value for no Jenkins connection.
The ros_buildfarm pipeline has branching behavior whenever being run on
a workstation, in which case it connects to Jenkins via HTTP, and in a
build farm job , in which case it emits Groovy scripts to make the
equivalent changes.

In order to avoid hitting heavyweight API endpoints for presence
queries, always check whether the jenkins API client is `None` rather
than different paths checking for for it to be False or None.

Previously these false-y values were interchangable but not in this
formulation.

Another alternative would be to make all checks:
`not isinstance(Jenkins)`
which would keep the versatility without trying to check the truthiness,
and thus length of the Jenkins object, avoiding the heavyweight API
call.