Use sftp via shell to publish package metadata files.
The publish-over-ssh plugin adds an extra configuration step that's hard
to manage securely as it does not use the Jenkins ssh-agent and
credential system, nor does it support openssh formatted or ed22519
keys, unlike the Jenkins private key credential.

Although it requires a bit more finagling ourselves, I think this is an
equally valid way to publish this content which only requires one copy
of the ssh private key in the ssh-agent and not an out-of-band system.

I chose sftp over rsync in order to batch the transfers of these smaller
files without doing wildly long rsync commands.