Additional fixes for RPM triggers for upload jobs
It turns out that the format of the sync-to-testing job names is even
more complicated than what was constructed here, so we were still adding
incorrect job names there. The previous update to this function didn't
regress that, but it also didn't fix it.

It turns out that there are functions to just do the formatting for us,
so we should just use that.