Restore PM-specific option specification in mock config
It appears that we actually WANT this configuration key to change with
the package manager, and this will be required as mock now supports
multiple DNF tools.

This is all a bit of a mess right now. There isn't a straightforward way
to determine which package manager we should be using, and mock is in
the process of transitioning the default PM to dnf5, where it was
previously unversioned dnf (which means dnf4).

This reverts part of ed19495fa10107f32d5f2cd29e5e5f28bef4a622