Use 'host' network in deb container builds (#1071)
The --network argument was added to 'docker build' in Docker 1.13 in
2017 and behaves like the --net argument to 'docker run'.

We're already using --net=host during the actual deb build and during
all of the RPM containers. The change here is to also use host
networking during the build phase.

The intent here is to support more self-contained building scenarios
where a repository may be on localhost.