Deprecate 'options' argument to 'expand_template'
The only known use of the 'options' argument is to force EmPy to treat
bangpaths as comments. This change promotes that option to an explicit
keyword argument of 'expand_template', which decouples the API from the
EmPy API.

Any current uses of 'options' will continue to work, and the new
'ignore_bangpath' option will take precedence in a collision.