Add support for EmPy 4 bangpath suppression mechanism
The previous BANGPATH_OPT value expands to 'processBangpaths' in empy3
which is the inverse of the default `ignoreBangpaths` configuration
option in empy4.

It's unfortunate that these options are reversed since it complicates
our handling of 3 and 4.

I am not sure if this is cleaner than introspecting on the empy version
and sending the appropriate configuration but I'll start here and
iterate on review.

Co-authored-by: Scott K Logan <logans@cottsay.net>